# contracts/generated/api-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "BriefBindingV1": {
        "additionalProperties": false,
        "properties": {
          "access_policy": {
            "$ref": "#/components/schemas/DataAccess"
          },
          "dataset_revision_id": {
            "$ref": "#/components/schemas/Id"
          },
          "role": {
            "$ref": "#/components/schemas/DataPartition"
          }
        },
        "required": [
          "dataset_revision_id",
          "role",
          "access_policy"
        ],
        "type": "object"
      },
...
      "BriefCreate": {
        "additionalProperties": false,
        "properties": {
          "bindings": {
            "items": {
              "$ref": "#/components/schemas/BriefBindingV1"
            },
            "maxItems": 64,
            "minItems": 1,
            "type": "array"
          },
          "content": {
            "$ref": "#/components/schemas/BriefContentV1"
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          },
          "supersedes_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Id"
              }
            ]
          }
        },
        "required": [
          "schema_version",
          "content",
          "bindings"
        ],
        "type": "object"
      },
...
      "BriefState": {
        "enum": [
          "DRAFT",
          "FROZEN"
        ],
        "type": "string"
      },
      "BriefUpdate": {
        "additionalProperties": false,
        "properties": {
          "bindings": {
            "items": {
              "$ref": "#/components/schemas/BriefBindingV1"
            },
            "maxItems": 64,
            "minItems": 1,
            "type": "array"
          },
          "content": {
            "$ref": "#/components/schemas/BriefContentV1"
          },
          "expected_revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          }
        },
        "required": [
          "schema_version",
          "expected_revision",
          "content",
          "bindings"
        ],
        "type": "object"
      },
      "BriefView": {
        "additionalProperties": false,
        "properties": {
          "bindings": {
            "items": {
              "$ref": "#/components/schemas/BriefBindingV1"
            },
            "maxItems": 64,
            "minItems": 1,
            "type": "array"
          },
          "content": {
            "$ref": "#/components/schemas/BriefContentV1"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "frozen_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "project_id": {
            "$ref": "#/components/schemas/Id"
          },
          "revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "state": {
            "$ref": "#/components/schemas/BriefState"
          },
          "supersedes_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Id"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int32",
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "project_id",
          "version",
          "revision",
          "state",
          "content",
          "bindings",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "CommandResult_BriefView": {
        "additionalProperties": false,
        "properties": {
          "replayed": {
            "type": "boolean"
          },
          "resource": {
            "additionalProperties": false,
            "properties": {
              "bindings": {
                "items": {
                  "$ref": "#/components/schemas/BriefBindingV1"
                },
                "maxItems": 64,
                "minItems": 1,
                "type": "array"
              },
              "content": {
                "$ref": "#/components/schemas/BriefContentV1"
              },
              "created_at": {
                "format": "date-time",
                "type": "string"
              },
              "frozen_at": {
                "format": "date-time",
                "type": [
                  "string",
                  "null"
                ]
              },
              "id": {
                "$ref": "#/components/schemas/Id"
              },
              "project_id": {
                "$ref": "#/components/schemas/Id"
              },
              "revision": {
                "$ref": "#/components/schemas/Revision"
              },
              "state": {
                "$ref": "#/components/schemas/BriefState"
              },
              "supersedes_id": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/Id"
                  }
                ]
              },
              "updated_at": {
                "format": "date-time",
                "type": "string"
              },
              "version": {
                "format": "int32",
                "maximum": 2147483647,
                "minimum": 1,
                "type": "integer"
              }
            },
            "required": [
              "id",
              "project_id",
              "version",
              "revision",
              "state",
              "content",
              "bindings",
              "created_at",
              "updated_at"
            ],
            "type": "object"
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          }
        },
        "required": [
          "schema_version",
          "replayed",
          "resource"
        ],
        "type": "object"
      },
...
      "CommandResult_ProjectView": {
        "additionalProperties": false,
        "properties": {
          "replayed": {
            "type": "boolean"
          },
          "resource": {
            "additionalProperties": false,
            "properties": {
              "archived_at": {
                "format": "date-time",
                "type": [
                  "string",
                  "null"
                ]
              },
              "created_at": {
                "format": "date-time",
                "type": "string"
              },
              "created_by": {
                "$ref": "#/components/schemas/ProjectOrigin"
              },
              "current_automation_policy_id": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/Id"
                  }
                ]
              },
              "current_brief_id": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/Id"
                  }
                ]
              },
              "description": {
                "type": "string"
              },
              "id": {
                "$ref": "#/components/schemas/Id"
              },
              "name": {
                "type": "string"
              },
              "revision": {
                "$ref": "#/components/schemas/Revision"
              },
              "root_lineage_id": {
                "$ref": "#/components/schemas/Id"
              },
              "state": {
                "$ref": "#/components/schemas/ProjectState"
              },
              "updated_at": {
                "format": "date-time",
                "type": "string"
              }
            },
            "required": [
              "id",
              "root_lineage_id",
              "name",
              "description",
              "state",
              "created_by",
              "created_at",
              "updated_at",
              "revision"
            ],
            "type": "object"
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          }
        },
        "required": [
          "schema_version",
          "replayed",
          "resource"
        ],
        "type": "object"
      },
...
      "Id": {
        "format": "uuid",
        "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
        "type": "string"
      },
...
      "Page_ProjectView": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "archived_at": {
                  "format": "date-time",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "created_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "created_by": {
                  "$ref": "#/components/schemas/ProjectOrigin"
                },
                "current_automation_policy_id": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/Id"
                    }
                  ]
                },
                "current_brief_id": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/Id"
                    }
                  ]
                },
                "description": {
                  "type": "string"
                },
                "id": {
                  "$ref": "#/components/schemas/Id"
                },
                "name": {
                  "type": "string"
                },
                "revision": {
                  "$ref": "#/components/schemas/Revision"
                },
                "root_lineage_id": {
                  "$ref": "#/components/schemas/Id"
                },
                "state": {
                  "$ref": "#/components/schemas/ProjectState"
                },
                "updated_at": {
                  "format": "date-time",
                  "type": "string"
                }
              },
              "required": [
                "id",
                "root_lineage_id",
                "name",
                "description",
                "state",
                "created_by",
                "created_at",
                "updated_at",
                "revision"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "next_cursor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Id"
              }
            ]
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          }
        },
        "required": [
          "schema_version",
          "items"
        ],
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "type": "string"
          },
          "current_revision": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Revision"
              }
            ]
          },
          "detail": {
            "type": "string"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "request_id": {
            "$ref": "#/components/schemas/Id"
          },
          "retryable": {
            "type": "boolean"
          },
          "safe_next_actions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "status": {
            "format": "int32",
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "request_id",
          "retryable",
          "field_errors",
          "safe_next_actions"
        ],
        "type": "object"
      },
      "ProjectCreate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 8000,
            "type": "string"
          },
          "fork_from_project_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Id"
              }
            ]
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "schema_version": {
            "$ref": "#/components/schemas/SchemaV1"
          }
        },
        "required": [
          "schema_version",
          "name",
          "description"
        ],
        "type": "object"
      },
...
      "Revision": {
        "description": "Canonical decimal string in the PostgreSQL signed bigint range; nonnegative counters or positive revisions.",
        "maxLength": 19,
        "minLength": 1,
        "pattern": "^(?:[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6][0-9]{0}|9223372036854775807)(?![\\s\\S])",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "BrowserSession": {
        "description": "Native private cookie; browser writes require exact same-origin Origin. Explicit loopback development uses quazonai-dev.",
        "in": "cookie",
        "name": "__Host-quazonai",
        "type": "apiKey"
      },
      "MachineBearer": {
        "bearerFormat": "qz2.UUIDv7.opaque-capability",
        "description": "Opaque native capability; only project/run/downstream-scoped server records confer authority. Never combine with browser Cookie.",
        "scheme": "bearer",
        "type": "http"
      },
      "OperatorCommandGrant": {
        "description": "One-time TOTP-verified CLI grant bound to this credential, exact operation, target and full nonsecret request. No Agent/automation grant issuance.",
        "in": "header",
        "name": "X-Operator-Grant",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "QuaZonai HTTP API",
    "version": "2.0.0-dev.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/briefs/{id}": {
      "get": {
        "operationId": "get_brief",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefView"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Research briefs"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult_BriefView"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": [],
            "OperatorCommandGrant": []
          }
        ],
        "tags": [
          "Research briefs"
        ]
      }
    },
...
    "/api/v2/projects": {
      "get": {
        "operationId": "projects",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectView"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "operationId": "create_project",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult_ProjectView"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": [],
            "OperatorCommandGrant": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/api/v2/projects/{id}/briefs": {
      "get": {
        "operationId": "list_briefs",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BriefView"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": []
          }
        ],
        "tags": [
          "Research briefs"
        ]
      },
      "post": {
        "operationId": "create_brief",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult_BriefView"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Shared native machine attempt limit or bounded crypto capacity; respect Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "BrowserSession": []
          },
          {
            "MachineBearer": [],
            "OperatorCommandGrant": []
          }
        ],
        "tags": [
          "Research briefs"
        ]
      }
    },
...
  },
...
}